requirements.txt
pyserial>=3.5
fastapi>=0.109
uvicorn[standard]>=0.27
websockets>=12.0
aiortc>=1.6
av>=11.0
requests>=2.31
httpx>=0.26
Pillow>=10.0
opencv-python-headless>=4.8,<4.11
face_recognition>=1.3
deepface>=0.0.79
tf-keras>=2.16
numpy>=1.24,<2
pymongo[srv]
python-dotenv
elevenlabs
simple-ai-agents>=0.5
openai>=1.0
google-generativeai>=0.8